+++ crypto-sodium/lib/Crypto/Key.hs
@@ -6,12 +6,86 @@
 --
 -- SPDX-License-Identifier: MPL-2.0
 
--- | Utilities for working with secret keys.
+-- | This module gives different ways of obtaining secret keys.
+--
+-- = Key derivation (from a password)
+--
+-- Sometimes, instead of generating fresh random transient encryption keys,
+-- you want an encryption key to be persistent and you don’t want to
+-- store it anywhere – instead you ask the user to provide it.
+-- Such a secret value known and entered by the user is usually called
+-- a “password”.
+--
+-- However, passwords make terrible encryption keys because encryption keys:
+--
+-- 1. often need to have a specific exact length and
+-- 2. need to be hard to guess (or brute-force).
+--
+-- Item 1 above can be easily ticked off by deriving the encryption key from
+-- the password by applying a hash-function to it, however in order to
+-- achieve 2 we need our “hash-function” to:
+--
+-- * be slow to compute (to make brute-forcing less feasible) and
+-- * mix extra “noise” into the derivation process to make it harder
+--   to pre-compute derived values in advance.
+--
+-- A construction that satisfies both requirements is called a /key derivation
+-- function (KDF)/. This module provides a convenient interface for deriving
+-- secure keys from passwords by the way of one such KDF.
+--
+-- == Use
+--
+-- This module provides two functions: 'derive' and 'rederive'.
+-- You can think of the entire process similar to how you set the password
+-- on your account at some website once, and then use this password to log in.
+--
+-- When you derive a key for the first time (e.g. you ask the user to enter their
+-- password twice, and then encrypt something), you use the 'derive' function,
+-- which gives you the derived key and a /derivation slip/. The slip is not
+-- secret, you can store it in plain text and, in fact, you /have to/ store it
+-- in plaintext somewhere next to the encrypted data.
+--
+-- When you need to derive the key in the future (e.g. to decrypt some previously
+-- encrypted data), you will need the user’s password (ask them) /and/ you
+-- will need the original derivation slip, which you should have stored.
+-- You pass these to 'rederive' and it will give you the same key.
+--
+-- @
+-- import qualified Crypto.Key as Key
+--
+-- encrypt = do
+--   password <- {- ask the user to enter their password -}
+--   password2 <- {- ask the user to confirm their password -}
+--   when (password /= password2 then) $ throwIO {- passwords do not match -}
+--
+--   let params = {- choose key derivation parameters -}
+--   (key, slip) <- Key.derive params password
+--
+--   {- store slip (it is not secret) -}
+--   {- encrypt data with key -}
+--
+-- decrypt = do
+--   password <- {- ask the user to enter their password -}
+--   slip <- {- get the stored slip -}
+--
+--   key <- Key.rederive slip password
+--
+--   {- decrypt data with key -}
+-- @
+--
+-- = Random key generation
+--
+-- The 'random' function is great at generating new secure secret keys.
 module Crypto.Key
-  ( fromPassword
-  , Params (..)
-  , type (!>=!)
+  ( type (!>=!)
 
+  -- * Key derivation
+  , Params (..)
+  , DerivationSlip
+  , derive
+  , rederive
+
+  -- * Random key generation
   , generate
   ) where
 
@@ -21,10 +95,11 @@ import Data.Kind (Constraint)
 import GHC.TypeLits (type (<=), KnownNat)
 import System.IO.Unsafe (unsafeDupablePerformIO)
 
-import Crypto.Pwhash.Internal (Algorithm (Argon2id_1_3), Params (..), Salt, pwhash)
-
 import qualified Libsodium as Na
 
+import Crypto.Key.Internal (DerivationSlip, Params (..))
+
+import qualified Crypto.Key.Internal as I
 import qualified Crypto.Random
 
 
@@ -50,33 +125,64 @@ type family a !>=! b :: Constraint where
   a !>=! b = ()
 class LessSecureStorage a b
 
--- | Derive an encryption key from a password using a secure KDF.
+
+-- | Derive a key from a password using a secure KDF for the first time.
 --
--- __Warning: this function is can be tricky to use and therefore it is
--- likely to change in the future versions of the library. In the meantime,
--- carefully read and closely follow the recommendations below.__
+-- This function takes two arguments:
 --
--- The purpose of the nonce is to protect against the user choosing a weak
--- password and the attacker being able to precompute the derived secret key.
--- For this reason, the recommended strategy is to generate a new random
--- nonce when deriving the encryption key for the first time and then store
--- the nonce for deriving the key in the future (e.g. for decryption).
--- The nonce is not secret and can be stored as plaintext.
+-- 1. key derivation parameters, which specify how slow the derivation process
+-- will be (the slower you can afford the better for security),
+--
+-- 2. the user’s password to derive the key from.
 --
 -- See @libsodium@ documentation for how to determine 'Params'.
--- It is strongly recommended to store the 'Params' used for derivation
--- alongside the encrypted data instead of relying on code constants.
-fromPassword
-  ::  ( ByteArrayAccess passwd, ByteArrayAccess nonceBytes
-      , ByteArrayN n hash, hash !>=! passwd
+--
+-- It returns the derived key and a /slip/ that you need to save in order to be
+-- able to derive the same key from the same password in the future. The slip
+-- is not secret, so you can store it in plaintext; just make sure you can
+-- access it in the future, as you will need to provide it to 'rederive'.
+--
+-- It can derive a key of almost any length and the output length is encoded
+-- in the type. There is an additional type-level restriction which forces
+-- you to store the derived key in memory at least as securely as you
+-- stored the password.
+--
+-- Note: This function is not thread-safe until Sodium is initialised.
+-- See "Crypto.Init" for details.
+derive
+  ::  forall key n passwd.
+      ( ByteArrayAccess passwd
+      , ByteArrayN n key, key !>=! passwd
       , Na.CRYPTO_PWHASH_BYTES_MIN <= n, n <= Na.CRYPTO_PWHASH_BYTES_MAX
       )
-  => Params -- ^ Hashing parameters.
-  -> passwd  -- ^ Password to hash.
-  -> Salt nonceBytes  -- ^ Nonce used for deriving the key from the password.
-  -> Maybe hash
-fromPassword params passwd salt =
-  unsafeDupablePerformIO $ pwhash Argon2id_1_3 params passwd salt
+  => I.Params -- ^ Derivation parameters.
+  -> passwd  -- ^ Password to derive from.
+  -> IO (Maybe (key, I.DerivationSlip))
+derive = I.derive
+
+-- | Reerive a key from a password using a secure KDF.
+--
+-- This function takes two arguments:
+--
+-- 1. A derivation slip previously returned by 'derive'.
+--
+-- 2. The user’s password.
+--
+-- This function is guaranteed to derive the same key from the same password
+-- as long as the same derivation slip was provided.
+--
+-- See 'derive' for additional details.
+rederive
+  ::  forall key n passwd.
+      ( ByteArrayAccess passwd
+      , ByteArrayN n key, key !>=! passwd
+      , Na.CRYPTO_PWHASH_BYTES_MIN <= n, n <= Na.CRYPTO_PWHASH_BYTES_MAX
+      )
+  => I.DerivationSlip -- ^ Original derivation slip.
+  -> passwd  -- ^ Password to rederive from.
+  -> Maybe key
+rederive slip passwd =
+  unsafeDupablePerformIO $ I.rederive slip passwd
   -- This IO is safe, because it is pure.