+++ crypto-sodium/CHANGELOG.md
@@ -6,4 +6,4 @@
 
 * `sodiumInit`
 * Random bytes generation: `Crypto.Random`
-* Key derivation: `Crypto.Key.fromPassword`
+* Key derivation: `Crypto.Key.derive` and `Crypto.Key.rederive`